questionsを外に括りだす
