+++ src/api/app.py
@@ -43,99 +43,18 @@ def before_request():
 def teardown_request(exception):
     pass
 
-# マッパークラス TODO 外に出すこと
-## Mapper For Operation
-class OperationMapper(Mapper):
-    id = RawField()
-    username = RawField()
-    state = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-class ListOperationMapper(Mapper):
-    result = ListDelegateField(OperationMapper)
-
-## Mapper For Answer
-class AnswerMapper(Mapper):
-    id = RawField()
-    question_id = RawField()
-    content = RawField()
-    state = RawField()
-    created_by = RawField()
-    updated_by = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-class ListAnswerMapper(Mapper):
-    pass
-
-## Mapper For Question
-class QuestionMapper(Mapper):
-    id = RawField()
-    content = RawField()
-    state = RawField()
-    created_by = RawField()
-    updated_by = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-    answers = ListDelegateField(AnswerMapper)
-
-class ListQuestionMapper(Mapper):
-    #question_list = ListDelegateField(QuestionMapper)
-    result = ListDelegateField(QuestionMapper)
-
-
-## Mapper For Inofrmation
-class InformationMapper(Mapper):
-    id = RawField()
-    content = RawField()
-    state = RawField()
-    created_by = RawField()
-    updated_by = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-class ListInformationMapper(Mapper):
-    result = ListDelegateField(InformationMapper)
-
-## Mapper For Tweet
-class TweetMapper(Mapper):
-    id = RawField()
-    type = RawField()
-    tweet_id = RawField()
-    content = RawField()
-    created_by = RawField()
-    updated_by = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-class ListTweetMapper(Mapper):
-    result = ListDelegateField(TweetMapper)
-
-## Mapper For Response
-class ResponseMapper(Mapper):
-    id = RawField()
-    type = RawField()
-    content = RawField()
-    state = RawField()
-    created_by = RawField()
-    updated_by = RawField()
-    created_at = RawField()
-    updated_at = RawField()
-
-class ListResponseMapper(Mapper):
-    result = ListDelegateField(ResponseMapper)
 
 def clear_session():
     session.clear()
     pass
+from questions import questions
+app.register_blueprint(questions.app, url_prefix="/questions")
 
-@app.route('/')
-@crossdomain(origin='*')
-def test():
-    res = {1: 2}
-    return jsonify(res)
+#@app.route('/')
+#@crossdomain(origin='*')
+#def test():
+#    res = {1: 2}
+#    return jsonify(res)
 
 @app.route('/operations', methods=['POST', 'OPTIONS'])
 @crossdomain(origin='*')
@@ -256,153 +175,153 @@ def delete_operation(operation_id):
 
     return jsonify(status_code=code)
 
-@app.route('/questions', methods=['GET'])
-@crossdomain(origin='*')
-def index_questions():
-    """質問一覧を返却します
-    """
-    code = 200
-    #logging.debug(request.headers)
-    # TODO 公開時にコメントイン
-    #if request.headers['Api-Key'] != API_ACCESS_KEY:
-    #    abort(401)
-
-    questions_dict = {}
-    try:
-        questions = get_questions()
-        logging.debug(questions)
-        questions_dict = ListQuestionMapper({'result': questions}).as_dict()
-    except:
-        pass
-    logging.debug(questions_dict)
-    result = questions_dict['result']
-
-    return jsonify(status_code=code, result=result)
-
-# TODO answerも同時に登録するように修正すること
-@app.route('/questions', methods=['POST'])
-@crossdomain(origin='*')
-def add_question():
-    """リクエストを元に質問を登録します
-    """
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    req = request.form
-    #logging.debug(req['content'])
-    #logging.debug(req['state'])
-    #logging.debug(req['created_by'])
-    #logging.debug(req['updated_by'])
-    try:
-        question = Question(
-                            id=None,
-                            content=req['content'],
-                            state=req['state'],
-                            created_by=req['created_by'],
-                            updated_by=req['updated_by'],
-                            created_at=tstr,
-                            updated_at=tstr
-        )
-        db_session.add(question)
-        db_session.commit()
-    except:
-        # 登録失敗
-        db_session.rollback()
-    finally:
-        db_session.close()
-
-    return jsonify(status_code=code)
-
-@app.route('/questions/<question_id>', methods=['GET'])
-@crossdomain(origin='*')
-def show_question(question_id):
-    """idを元に質問データを取得します
-    """
-    code = 200
-    question_dict = {}
-
-    try:
-        question = get_question(question_id)
-        question_dict = QuestionMapper(question).as_dict()
-    except:
-        # 取得に失敗
-        pass
-
-    return jsonify(status_code=code, result=question_dict)
-
-def get_question(question_id):
-    """質問idを元にQuestionモデルを返却する
-    """
-    #logging.debug(Question.query.first().id)
-    question = []
-    res = Question.query.filter("id = :question_id").params(question_id=question_id).first()
-    #logging.debug(res)
-    question = Question(id=question_id,
-                        content=res.content,
-                        state=res.state,
-                        created_by=res.created_by,
-                        updated_by=res.updated_by,
-                        created_at=res.created_at,
-                        updated_at=res.updated_at
-    )
-
-    return question
-
-def get_questions():
-    """質問一覧のリストを返却する
-    """
-    questions = []
-    res = Question.query.all()
-    for row in res:
-        # get_question(row["id"])でmodelをセットしていく？
-        questions.append(row)
-
-    return questions
-
-@app.route('/questions/<question_id>', methods=['PUT'])
-@crossdomain(origin='*')
-def edit_question(question_id):
-    """質問を更新します
-    """
-    code = 201
-    tdatetime = dt.now()
-    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
-    req = request.form
-    try:
-        row = db_session.query(Question).get(question_id)
-        row.content = req['content']
-        row.state = req['state']
-        row.created_by = req['created_by']
-        row.updated_by = req['updated_by']
-        row.updated_at = tstr
-        db_session.flush()
-        #db_session.commit()
-    except:
-        pass
-    finally:
-        pass
-
-    # app.logger.debug(request.form)
-    return jsonify(status_code=code)
-
-@app.route('/questions/<question_id>', methods=['DELETE'])
-@crossdomain(origin='*')
-def delete_question(question_id):
-    """質問を削除します
-    """
-    code = 204
-    try:
-        row = Question.query.get(question_id)
-        db_session.delete(row)
-        db_session.flush()
-    except:
-        pass
-    finally:
-        pass
-
-    #logging.debug(row.created_by)
-
-    return jsonify(status_code=code)
+#@app.route('/questions', methods=['GET'])
+#@crossdomain(origin='*')
+#def index_questions():
+#    """質問一覧を返却します
+#    """
+#    code = 200
+#    #logging.debug(request.headers)
+#    # TODO 公開時にコメントイン
+#    #if request.headers['Api-Key'] != API_ACCESS_KEY:
+#    #    abort(401)
+#
+#    questions_dict = {}
+#    try:
+#        questions = get_questions()
+#        logging.debug(questions)
+#        questions_dict = ListQuestionMapper({'result': questions}).as_dict()
+#    except:
+#        pass
+#    logging.debug(questions_dict)
+#    result = questions_dict['result']
+#
+#    return jsonify(status_code=code, result=result)
+#
+## TODO answerも同時に登録するように修正すること
+#@app.route('/questions', methods=['POST'])
+#@crossdomain(origin='*')
+#def add_question():
+#    """リクエストを元に質問を登録します
+#    """
+#    code = 201
+#    tdatetime = dt.now()
+#    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
+#    req = request.form
+#    #logging.debug(req['content'])
+#    #logging.debug(req['state'])
+#    #logging.debug(req['created_by'])
+#    #logging.debug(req['updated_by'])
+#    try:
+#        question = Question(
+#                            id=None,
+#                            content=req['content'],
+#                            state=req['state'],
+#                            created_by=req['created_by'],
+#                            updated_by=req['updated_by'],
+#                            created_at=tstr,
+#                            updated_at=tstr
+#        )
+#        db_session.add(question)
+#        db_session.commit()
+#    except:
+#        # 登録失敗
+#        db_session.rollback()
+#    finally:
+#        db_session.close()
+#
+#    return jsonify(status_code=code)
+#
+#@app.route('/questions/<question_id>', methods=['GET'])
+#@crossdomain(origin='*')
+#def show_question(question_id):
+#    """idを元に質問データを取得します
+#    """
+#    code = 200
+#    question_dict = {}
+#
+#    try:
+#        question = get_question(question_id)
+#        question_dict = QuestionMapper(question).as_dict()
+#    except:
+#        # 取得に失敗
+#        pass
+#
+#    return jsonify(status_code=code, result=question_dict)
+#
+#def get_question(question_id):
+#    """質問idを元にQuestionモデルを返却する
+#    """
+#    #logging.debug(Question.query.first().id)
+#    question = []
+#    res = Question.query.filter("id = :question_id").params(question_id=question_id).first()
+#    #logging.debug(res)
+#    question = Question(id=question_id,
+#                        content=res.content,
+#                        state=res.state,
+#                        created_by=res.created_by,
+#                        updated_by=res.updated_by,
+#                        created_at=res.created_at,
+#                        updated_at=res.updated_at
+#    )
+#
+#    return question
+#
+#def get_questions():
+#    """質問一覧のリストを返却する
+#    """
+#    questions = []
+#    res = Question.query.all()
+#    for row in res:
+#        # get_question(row["id"])でmodelをセットしていく？
+#        questions.append(row)
+#
+#    return questions
+#
+#@app.route('/questions/<question_id>', methods=['PUT'])
+#@crossdomain(origin='*')
+#def edit_question(question_id):
+#    """質問を更新します
+#    """
+#    code = 201
+#    tdatetime = dt.now()
+#    tstr = tdatetime.strftime('%Y-%m-%d %H:%M:%S')
+#    req = request.form
+#    try:
+#        row = db_session.query(Question).get(question_id)
+#        row.content = req['content']
+#        row.state = req['state']
+#        row.created_by = req['created_by']
+#        row.updated_by = req['updated_by']
+#        row.updated_at = tstr
+#        db_session.flush()
+#        #db_session.commit()
+#    except:
+#        pass
+#    finally:
+#        pass
+#
+#    # app.logger.debug(request.form)
+#    return jsonify(status_code=code)
+#
+#@app.route('/questions/<question_id>', methods=['DELETE'])
+#@crossdomain(origin='*')
+#def delete_question(question_id):
+#    """質問を削除します
+#    """
+#    code = 204
+#    try:
+#        row = Question.query.get(question_id)
+#        db_session.delete(row)
+#        db_session.flush()
+#    except:
+#        pass
+#    finally:
+#        pass
+#
+#    #logging.debug(row.created_by)
+#
+#    return jsonify(status_code=code)
 
 
 # /answers
